Implement new schema for amphiro b1 data
Renamed several methods in repositories and updated the controllers.
Added comments.
Fixed code formatting (replaced tabs with spaces).
Configured admin and user controller actions to access data from the utilities to which the authenticated user has permissions.
Updated job that updates amphiro b1 data schema.
Implemented new schema for amphiro b1 data.
Created new HBase repository for amphiro b1 data.
Updated job for exporting trial data.

+++ src/main/java/eu/daiad/web/model/amphiro/AmphiroDataPoint.java
@@ -4,7 +4,7 @@ public class AmphiroDataPoint extends AmphiroAbstractDataPoint {
 
 	private long sessionId;
 
-	private int index;
+	private long index;
 
 	private boolean history;
 
@@ -16,11 +16,11 @@ public class AmphiroDataPoint extends AmphiroAbstractDataPoint {
 		this.sessionId = sessionId;
 	}
 
-	public int getIndex() {
+	public long getIndex() {
 		return index;
 	}
 
-	public void setIndex(int index) {
+	public void setIndex(long index) {
 		this.index = index;
 	}
 
@@ -31,7 +31,7 @@ public class AmphiroDataPoint extends AmphiroAbstractDataPoint {
 	public void setHistory(boolean history) {
 		this.history = history;
 	}
-	
+
 	public void setTimestamp(long timestamp) {
 		this.timestamp = timestamp;
 	}
